add drink controller, repository, service

+++ src/main/kotlin/hu/havasig/alcoholCalendar/model/Drink.kt
@@ -8,7 +8,7 @@ import javax.persistence.*
 @Table(name = "drinks")
 class Drink(
 	var name: String = "",
-	var percentage: Int? = null,
+	var percentage: Double? = null,
 	var amount: Double? = null,
 	var date: LocalDate? = null
 ) {
